requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.6
aiofiles>=23.1.0
python-multipart>=0.0.6
apscheduler>=3.9.1
python-dotenv>=1.0.1
boto3>=1.35.0
APScheduler>=3.10.4
alembic>=1.13.2
